remove test limitations
